requirements.txt
discord.py
requests
paramiko
asyncio
requests
python-dotenv
Flask